+++ temba/channels/models.py
@@ -1,5 +1,4 @@
 
-import json
 import logging
 import time
 from abc import ABCMeta
@@ -33,7 +32,7 @@ from django.utils.http import urlquote_plus
 from django.utils.translation import ugettext_lazy as _
 
 from temba.orgs.models import NEXMO_APP_ID, NEXMO_APP_PRIVATE_KEY, NEXMO_KEY, NEXMO_SECRET, Org
-from temba.utils import analytics, dict_to_json, dict_to_struct, get_anonymous_user, on_transaction_commit
+from temba.utils import analytics, dict_to_struct, get_anonymous_user, json, on_transaction_commit
 from temba.utils.email import send_template_email
 from temba.utils.gsm7 import calculate_num_segments
 from temba.utils.models import JSONAsTextField, SquashableModel, TembaModel, generate_uuid
@@ -931,7 +930,7 @@ class Channel(TembaModel):
                 return None
             else:
                 cached = channel.as_cached_json()
-                cache.set(key, dict_to_json(cached), 900)
+                cache.set(key, json.dumps(cached), 900)
         else:
             cached = json.loads(cached)
 
@@ -1587,7 +1586,6 @@ class ChannelEvent(models.Model):
     TYPE_CALL_IN_MISSED = "mo_miss"
     TYPE_NEW_CONVERSATION = "new_conversation"
     TYPE_REFERRAL = "referral"
-    TYPE_FOLLOW = "follow"
     TYPE_STOP_CONTACT = "stop_contact"
 
     EXTRA_REFERRER_ID = "referrer_id"
@@ -1602,7 +1600,6 @@ class ChannelEvent(models.Model):
         (TYPE_STOP_CONTACT, _("Stop Contact"), "stop-contact"),
         (TYPE_NEW_CONVERSATION, _("New Conversation"), "new-conversation"),
         (TYPE_REFERRAL, _("Referral"), "referral"),
-        (TYPE_FOLLOW, _("Follow"), "follow"),
     )
 
     TYPE_CHOICES = [(t[0], t[1]) for t in TYPE_CONFIG]
@@ -1693,9 +1690,6 @@ class ChannelEvent(models.Model):
             handled = Trigger.catch_triggers(
                 self, Trigger.TYPE_REFERRAL, self.channel, referrer_id=self.extra.get("referrer_id"), extra=self.extra
             )
-
-        elif self.event_type == ChannelEvent.TYPE_FOLLOW:
-            handled = Trigger.catch_triggers(self, Trigger.TYPE_FOLLOW, self.channel)
 
         elif self.event_type == ChannelEvent.TYPE_STOP_CONTACT:
             user = get_anonymous_user()