+++ temba/flows/views.py
@@ -1,5 +1,4 @@
 
-import json
 import logging
 import traceback
 from datetime import datetime, timedelta
@@ -52,7 +51,7 @@ from temba.orgs.models import Org
 from temba.orgs.views import ModalMixin, OrgObjPermsMixin, OrgPermsMixin
 from temba.triggers.models import Trigger
 from temba.ussd.models import USSDSession
-from temba.utils import analytics, chunk_list, on_transaction_commit, str_to_bool
+from temba.utils import analytics, chunk_list, json, on_transaction_commit, str_to_bool
 from temba.utils.dates import datetime_to_ms, datetime_to_str
 from temba.utils.expressions import get_function_listing
 from temba.utils.s3 import public_file_storage
@@ -1378,7 +1377,7 @@ class FlowCRUDL(SmartCRUDL):
                 id = self.request.GET["id"]
 
                 modified_on = iso8601.parse_date(modified_on)
-                runs = runs.filter(modified_on__lte=modified_on).exclude(id__gte=id)
+                runs = runs.filter(modified_on__lte=modified_on).exclude(id=id)
 
             # we grab one more than our page to denote whether there's more to get
             runs = list(runs.order_by("-modified_on")[: self.paginate_by + 1])