+++ temba/utils/nexmo.py
@@ -1,4 +1,3 @@
-import json
 import time
 import uuid
 
@@ -7,6 +6,8 @@ import nexmo as nx
 import requests
 
 from django.utils.encoding import force_bytes
+
+from temba.utils import json
 
 
 class NexmoValidationError(Exception):
@@ -25,7 +26,7 @@ class NexmoClient(nx.Client):
         kwargs["key"] = api_key.strip()
         kwargs["secret"] = api_secret.strip()
         kwargs["application_id"] = app_id.strip()
-        kwargs["private_key"] = app_private_key.strip()
+        kwargs["private_key"] = app_private_key
         nx.Client.__init__(self, **kwargs)
 
     def update_account(self, mo_url, dr_url):  # pragma: needs cover